// project_phase3_sla_analytics/backend/src/generated/prisma/enums.ts
export const ScopeType = {
  SELF: 'SELF',
  TEAM: 'TEAM',
  DEPARTMENT: 'DEPARTMENT',
  DEPARTMENT_SUBTREE: 'DEPARTMENT_SUBTREE',
  RELATED_DEPARTMENTS: 'RELATED_DEPARTMENTS',
  ORG_WIDE: 'ORG_WIDE'
} as const

export type ScopeType = (typeof ScopeType)[keyof typeof ScopeType]
// ...
export const RepairType = {
  IN_HOUSE: 'IN_HOUSE',
  ON_SITE: 'ON_SITE'
} as const

export type RepairType = (typeof RepairType)[keyof typeof RepairType]


export const RepairStatus = {
  REGISTERED: 'REGISTERED',
  WAITING_REVIEW: 'WAITING_REVIEW',
  WAITING_COST_APPROVAL: 'WAITING_COST_APPROVAL',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  IN_REPAIR: 'IN_REPAIR',
  QC: 'QC',
  READY_FOR_DELIVERY: 'READY_FOR_DELIVERY',
  DELIVERED: 'DELIVERED',
  CLOSED: 'CLOSED',
  CANCELED: 'CANCELED',
  NO_REPAIR_REQUIRED: 'NO_REPAIR_REQUIRED'
} as const

export type RepairStatus = (typeof RepairStatus)[keyof typeof RepairStatus]